refactor: extract quiz types in a separate file
