+++ src/types/index.ts
@@ -1,3 +1,5 @@
-export type { AuthStateType, AuthContextType, AuthActionType, QuizContextType, GameContextType, QuizStateType, GameStateType } from "./contextTypes/auth.types";
+export type { AuthStateType, AuthContextType, AuthActionType } from "./contextTypes/auth.types";
+export type { QuizContextType, QuizStateType } from "./contextTypes/quiz.types";
+export type { GameContextType, QuizStateType } from "./contextTypes/game.types";
 export type { ReactChildrenType } from "./common.types";
 export type { ServicesType } from "./serviceTypes/service.types";
